# docs/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is HexLabs' API Documentation.",
    "title": "API Reference",
    "version": "1.0.0"
  },
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/files/upload": {
      "post": {
        "tags": ["files"],
        "summary": "Upload a new file.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "File successfully uploaded"
          },
          "400": {
            "description": "No file uploaded!"
          }
        }
      }
    },
    "/files/{id}": {
      "get": {
        "tags": ["files"],
        "summary": "Finds and displays file url by id.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File id that needs to be considered for filter.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": ""
              }
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
    "/files/{id}/download": {
      "get": {
        "tags": ["files"],
        "summary": "Finds and displays file download url by id.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File id that needs to be considered for filter.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": ""
              }
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
    "/hexathons": {
      "get": {
        "tags": ["hexathons"],
        "summary": "Finds and displays information of all hexathons.",
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Hexathon"
              }
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      },
      "post": {
        "tags": ["hexathons"],
        "summary": "Create a new hexathon.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "shortCode": {
                  "type": "string",
                  "required": true,
                  "unique": true
                },
                "isActive": {
                  "type": "boolean"
                },
                "startDate": {
                  "type": "string",
                  "format": "date-time",
                  "required": true
                },
                "endDate": {
                  "type": "string",
                  "format": "date-time",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Hexathon"
              }
            }
          },
          "400": {
            "description": "Please add all fields"
          }
        }
      }
    },
    "/hexathons/{id}": {
      "get": {
        "tags": ["hexathons"],
        "summary": "Finds and displays hexathon by id.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Hexathon id that needs to be considered for filter.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Hexathon"
              }
            }
          },
          "400": {
            "description": "Hexathon not found"
          }
        }
      },
      "put": {
        "tags": ["hexathons"],
        "summary": "Update an existing hexathon by id.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Hexathon id that needs to be considered for filter.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Hexathon"
              }
            }
          },
          "400": {
            "description": "Hexathon not found"
          }
        }
      }
    },
    "/interactions": {
      "get": {
        "tags": ["hexathons"],
        "summary": "Finds and displays information of all interactions.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "hexathon",
            "in": "query",
            "description": "hexathon id that needs to be considered for filter.",
            "required": true,
            "type": "integer",
            "collectionFormat": "multi"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "User id that needs to be considered for filter.",
            "required": true,
            "type": "integer",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Interaction"
              }
            }
          },
          "400": {
            "description": "Invalid status"
          }
        }
      },
      "post": {
        "tags": ["hexathons"],
        "summary": "Create a new interaction.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "identifier": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Interaction"
              }
            }
          },
          "400": {
            "description": "Interaction already exists for this user and identifier"
          }
        }
      }
    },
    "/interactions/{id}": {
      "put": {
        "tags": ["hexathons"],
        "summary": "Update an existing interaction by id.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Interaction id that needs to be considered for filter.",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "string"
                },
                "identifier": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Interaction"
              }
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
...
    "/statistics": {
      "get": {
        "tags": ["registration"],
        "summary": "Displays registration statistics.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "hexathon",
            "in": "query",
            "type": "integer",
            "description": "Hexathon ID for filter.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "userStatistics": {
                  "type": "object",
                  "properties": {
                    "totalUsers": {
                      "type": "integer"
                    },
                    "appliedUsers": {
                      "type": "integer"
                    },
                    "acceptedUsers": {
                      "type": "integer"
                    },
                    "confirmedUsers": {
                      "type": "integer"
                    },
                    "nonConfirmedUsers": {
                      "type": "integer"
                    },
                    "deniedUsers": {
                      "type": "integer"
                    }
                  }
                },
                "applicationStatistics": {
                  "type": "object"
                },
                "confirmationStatistics": {
                  "type": "object"
                },
                "rejectionStatistics": {
                  "type": "object"
                },
                "applicationDataStatistics": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Hexathon": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "shortCode": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean",
          "default": true
        },
        "startDate": {
          "type": "string",
          "format": "date-time"
        },
        "endDate": {
          "type": "string",
          "format": "date-time"
        }
      },
      "xml": {
        "name": "Hexathon"
      }
    },
    "Interaction": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": ["event", "scavengerHunt"]
        },
        "identifier": {
          "type": "string"
        },
        "timeStamp": {
          "type": "string",
          "format": "date-time"
        },
        "hexathon": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Interaction"
      }
    },
...
  }
}